changes to organize better classes
