+++ includes/Checker/Default_Check_Repository.php
@@ -74,7 +74,8 @@ class Default_Check_Repository extends Empty_Check_Repository {
 				'i18n_usage'                 => new Checks\General\I18n_Usage_Check(),
 				'enqueued_scripts_size'      => new Checks\Performance\Enqueued_Scripts_Size_Check(),
 				'enqueued_styles_size'       => new Checks\Performance\Enqueued_Styles_Size_Check(),
-				'code_obfuscation'           => new Checks\Plugin_Repo\Code_NotAllowed_Check(),
+				'code_obfuscation'           => new Checks\Plugin_Repo\Code_Obfuscation_Check(),
+				'plugin_content'             => new Checks\Plugin_Repo\Plugin_Content_Check(),
 				'file_type'                  => new Checks\Plugin_Repo\File_Type_Check(),
 				'plugin_header_fields'       => new Checks\Plugin_Repo\Plugin_Header_Fields_Check(),
 				'late_escaping'              => new Checks\Security\Late_Escaping_Check(),
