+++ tests/phpunit/tests/Checker/Checks/Code_Obfuscation_Check_Tests copy.php
@@ -1,15 +1,15 @@
 <?php
 /**
- * Tests for the Code_NotAllowed_Check class.
+ * Tests for the Code_Obfuscation_Check class.
  *
  * @package plugin-check
  */
 
 use WordPress\Plugin_Check\Checker\Check_Context;
 use WordPress\Plugin_Check\Checker\Check_Result;
-use WordPress\Plugin_Check\Checker\Checks\Plugin_Repo\Code_NotAllowed_Check;
+use WordPress\Plugin_Check\Checker\Checks\Plugin_Repo\Code_Obfuscation_Check;
 
-class Code_NotAllowed_Check_Tests extends WP_UnitTestCase {
+class Code_Obfuscation_Check_Tests extends WP_UnitTestCase {
 
 	/**
 	 * @dataProvider data_obfuscation_services
@@ -19,7 +19,7 @@ class Code_NotAllowed_Check_Tests extends WP_UnitTestCase {
 		$check_context = new Check_Context( UNIT_TESTS_PLUGIN_DIR . $plugin_basename );
 		$check_result  = new Check_Result( $check_context );
 
-		$check = new Code_NotAllowed_Check( $type_flag );
+		$check = new Code_Obfuscation_Check( $type_flag );
 		$check->run( $check_result );
 
 		$errors = $check_result->get_errors();
@@ -35,21 +35,21 @@ class Code_NotAllowed_Check_Tests extends WP_UnitTestCase {
 	public function data_obfuscation_services() {
 		return array(
 			'Zend Guard'      => array(
-				Code_NotAllowed_Check::TYPE_ZEND,
+				Code_Obfuscation_Check::TYPE_ZEND,
 				'test-plugin-code-obfuscation-zendguard-errors/load.php',
 				'obfuscated.php',
 				1,
 				1,
 			),
 			'Source Guardian' => array(
-				Code_NotAllowed_Check::TYPE_SOURCEGUARDIAN,
+				Code_Obfuscation_Check::TYPE_SOURCEGUARDIAN,
 				'test-plugin-code-obfuscation-sourceguardian-errors/load.php',
 				'obfuscated.php',
 				2,
 				4,
 			),
 			'ionCube'         => array(
-				Code_NotAllowed_Check::TYPE_IONCUBE,
+				Code_Obfuscation_Check::TYPE_IONCUBE,
 				'test-plugin-code-obfuscation-ioncube-errors/load.php',
 				'load.php',
 				16,
@@ -63,28 +63,12 @@ class Code_NotAllowed_Check_Tests extends WP_UnitTestCase {
 		$check_context = new Check_Context( UNIT_TESTS_PLUGIN_DIR . 'test-plugin-i18n-usage-without-errors/load.php' );
 		$check_result  = new Check_Result( $check_context );
 
-		$check = new Code_NotAllowed_Check();
+		$check = new Code_Obfuscation_Check();
 		$check->run( $check_result );
 
 		$errors = $check_result->get_errors();
 
 		$this->assertEmpty( $errors );
 		$this->assertSame( 0, $check_result->get_error_count() );
-	}
-
-	public function test_detect_five_stars_reviews_without_errors() {
-		$check_context = new Check_Context( UNIT_TESTS_PLUGIN_DIR . 'test-plugin-five-stars-without-errors/load.php' );
-		$check_result  = new Check_Result( $check_context );
-
-		$check = new Code_NotAllowed_Check();
-		$check->run( $check_result );
-
-		$errors = $check_result->get_errors();
-
-		$this->assertNotEmpty( $errors );
-		$this->assertSame( 1, $check_result->get_error_count() );
-
-		$this->assertTrue( isset( $errors['load.php'][16][61][0] ) );
-		$this->assertSame( 'five_star_reviews_detected', $errors['load.php'][16][61][0]['code'] );
 	}
 }
